requirements.txt
# Requirements file, to install them execute into
# virtual environment the command:
# $ pip install -r requirements.txt
jupyter
pandas
numpy
beautifulsoup4
requests
tabula-py